step to improving spec implementation

+++ api-spec/src/main/kotlin/eu/vendeli/rethis/api/spec/commands/cluster/ClusterKeySlotCommand.kt
@@ -7,6 +7,6 @@ import eu.vendeli.rethis.api.spec.common.types.RedisOperation
 import eu.vendeli.rethis.api.spec.common.types.RespCode
 
 @RedisCommand("CLUSTER KEYSLOT", RedisOperation.READ, [RespCode.INTEGER])
-fun interface ClusterKeySlotCommand : RedisCommandSpec<Int> {
+fun interface ClusterKeySlotCommand : RedisCommandSpec<Long> {
     suspend fun encode(key: String): CommandRequest<Nothing>
 }
